+++ components/camel-google/camel-google-calendar/src/test/java/org/apache/camel/component/google/calendar/CalendarColorsIT.java
@@ -14,11 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.google.calendar.integration;
+package org.apache.camel.component.google.calendar;
 
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.google.calendar.AbstractGoogleCalendarTestSupport;
-import org.apache.camel.component.google.calendar.internal.CalendarSettingsApiMethod;
+import org.apache.camel.component.google.calendar.internal.CalendarColorsApiMethod;
 import org.apache.camel.component.google.calendar.internal.GoogleCalendarApiCollection;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIf;
@@ -28,31 +27,22 @@ import org.slf4j.LoggerFactory;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 /**
- * Test class for {@link com.google.api.services.calendar.Calendar$Settings} APIs.
+ * Test class for {@link com.google.api.services.calendar.Calendar$Colors} APIs.
  */
 @EnabledIf(value = "org.apache.camel.component.google.calendar.AbstractGoogleCalendarTestSupport#hasCredentials",
            disabledReason = "Google Calendar credentials were not provided")
-public class CalendarSettingsIT extends AbstractGoogleCalendarTestSupport {
+public class CalendarColorsIT extends AbstractGoogleCalendarTestSupport {
 
-    private static final Logger LOG = LoggerFactory.getLogger(CalendarSettingsIT.class);
+    private static final Logger LOG = LoggerFactory.getLogger(CalendarColorsIT.class);
     private static final String PATH_PREFIX
-            = GoogleCalendarApiCollection.getCollection().getApiName(CalendarSettingsApiMethod.class).getName();
+            = GoogleCalendarApiCollection.getCollection().getApiName(CalendarColorsApiMethod.class).getName();
 
     @Test
     public void testGet() throws Exception {
-        // using String message body for single parameter "setting"
-        final com.google.api.services.calendar.model.Setting result = requestBody("direct://GET", "timezone");
+        com.google.api.services.calendar.model.Colors result = requestBody("direct://GET", null);
 
         assertNotNull(result, "get result");
         LOG.debug("get: " + result);
-    }
-
-    @Test
-    public void testList() throws Exception {
-        final com.google.api.services.calendar.model.Settings result = requestBody("direct://LIST", null);
-
-        assertNotNull(result, "list result");
-        LOG.debug("list: " + result);
     }
 
     @Override
@@ -61,13 +51,7 @@ public class CalendarSettingsIT extends AbstractGoogleCalendarTestSupport {
             @Override
             public void configure() {
                 // test route for get
-                from("direct://GET").to("google-calendar://" + PATH_PREFIX + "/get?inBody=setting");
-
-                // test route for list
-                from("direct://LIST").to("google-calendar://" + PATH_PREFIX + "/list");
-
-                // test route for watch
-                from("direct://WATCH").to("google-calendar://" + PATH_PREFIX + "/watch?inBody=contentChannel");
+                from("direct://GET").to("google-calendar://" + PATH_PREFIX + "/get");
 
             }
         };
