+++ components/camel-google/camel-google-drive/src/test/java/org/apache/camel/component/google/drive/DriveChangesIT.java
@@ -14,37 +14,52 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.google.drive.integration;
+package org.apache.camel.component.google.drive;
 
+import java.util.List;
+
+import com.google.api.services.drive.model.Change;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.google.drive.AbstractGoogleDriveTestSupport;
-import org.apache.camel.component.google.drive.internal.DriveAppsApiMethod;
+import org.apache.camel.component.google.drive.internal.DriveChangesApiMethod;
 import org.apache.camel.component.google.drive.internal.GoogleDriveApiCollection;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIf;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assumptions.assumeFalse;
 
 /**
- * Test class for com.google.api.services.drive.Drive$Apps APIs.
+ * Test class for com.google.api.services.drive.Drive$Changes APIs.
  */
 @EnabledIf(value = "org.apache.camel.component.google.drive.AbstractGoogleDriveTestSupport#hasCredentials",
            disabledReason = "Google Drive credentials were not provided")
-public class DriveAppsIT extends AbstractGoogleDriveTestSupport {
+public class DriveChangesIT extends AbstractGoogleDriveTestSupport {
 
-    private static final Logger LOG = LoggerFactory.getLogger(DriveAppsIT.class);
+    private static final Logger LOG = LoggerFactory.getLogger(DriveChangesIT.class);
     private static final String PATH_PREFIX
-            = GoogleDriveApiCollection.getCollection().getApiName(DriveAppsApiMethod.class).getName();
+            = GoogleDriveApiCollection.getCollection().getApiName(DriveChangesApiMethod.class).getName();
 
-    // TODO getting permission errors for this one
-    // May have to adjust scopes used
-    @Disabled
+    @Test
+    public void testGet() throws Exception {
+        final com.google.api.services.drive.model.ChangeList list = requestBody("direct://LIST", null);
+        List<Change> items = list.getItems();
+        assumeFalse(items.isEmpty());
+
+        Change change = items.get(0);
+        Long id = change.getId();
+
+        // using String message body for single parameter "changeId"
+        final com.google.api.services.drive.model.Change result = requestBody("direct://GET", id);
+
+        assertNotNull(result, "get result");
+        LOG.debug("get: " + result);
+    }
+
     @Test
     public void testList() throws Exception {
-        final com.google.api.services.drive.model.App result = requestBody("direct://LIST", null);
+        final com.google.api.services.drive.model.ChangeList result = requestBody("direct://LIST", null);
 
         assertNotNull(result, "list result");
         LOG.debug("list: " + result);
@@ -56,11 +71,15 @@ public class DriveAppsIT extends AbstractGoogleDriveTestSupport {
             public void configure() {
                 // test route for get
                 from("direct://GET")
-                        .to("google-drive://" + PATH_PREFIX + "/get?inBody=appId");
+                        .to("google-drive://" + PATH_PREFIX + "/get?inBody=changeId");
 
                 // test route for list
                 from("direct://LIST")
                         .to("google-drive://" + PATH_PREFIX + "/list");
+
+                // test route for watch
+                from("direct://WATCH")
+                        .to("google-drive://" + PATH_PREFIX + "/watch?inBody=contentChannel");
 
             }
         };
