+++ components/camel-google/camel-google-drive/src/test/java/org/apache/camel/component/google/drive/DriveChildrenIT.java
@@ -14,14 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.google.drive.integration;
+package org.apache.camel.component.google.drive;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import com.google.api.services.drive.model.File;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.google.drive.AbstractGoogleDriveTestSupport;
 import org.apache.camel.component.google.drive.internal.DriveChildrenApiMethod;
 import org.apache.camel.component.google.drive.internal.GoogleDriveApiCollection;
 import org.junit.jupiter.api.Test;
