+++ components/camel-google/camel-google-drive/src/test/java/org/apache/camel/component/google/drive/DriveCommentsIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.google.drive.integration;
+package org.apache.camel.component.google.drive;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -22,7 +22,6 @@ import java.util.Map;
 import com.google.api.services.drive.model.Comment;
 import com.google.api.services.drive.model.File;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.google.drive.AbstractGoogleDriveTestSupport;
 import org.apache.camel.component.google.drive.internal.DriveCommentsApiMethod;
 import org.apache.camel.component.google.drive.internal.DriveFilesApiMethod;
 import org.apache.camel.component.google.drive.internal.GoogleDriveApiCollection;
